+++ src/ModelIdentityMap.php
@@ -7,6 +7,7 @@ namespace Somnambulist\ReadModels;
 use IlluminateAgnostic\Str\Support\Str;
 use function array_intersect_key;
 use function array_key_exists;
+use function array_map;
 use function explode;
 use function func_get_args;
 use function get_class;
@@ -165,7 +166,7 @@ final class ModelIdentityMap
             return
                 $this->relationships[get_class($model)][(string)$identity][$related]
                 ??
-                $this->relationships[get_class($model)][$model->getPrimaryKey()][$related]
+                $this->relationships[get_class($model)][(string)$model->getPrimaryKey()][$related]
                 ??
                 []
             ;
@@ -174,7 +175,7 @@ final class ModelIdentityMap
         return
             $this->relationships[get_class($model)][(string)$identity]
             ??
-            $this->relationships[get_class($model)][$model->getPrimaryKey()]
+            $this->relationships[get_class($model)][(string)$model->getPrimaryKey()]
             ??
             []
         ;
@@ -185,7 +186,7 @@ final class ModelIdentityMap
         $class = get_class($model);
 
         if (!$this->has($class, (string)$id = $model->getPrimaryKey())) {
-            $this->identityMap[$class][$id] = $model;
+            $this->identityMap[$class][(string)$id] = $model;
 
             if (null !== $model->getExternalPrimaryKey()) {
                 $this->identityMap[$class][(string)$model->getExternalPrimaryKey()] = $model;
@@ -209,6 +210,8 @@ final class ModelIdentityMap
         if (!array_key_exists($class, $this->identityMap)) {
             return [];
         }
+
+        $ids = array_map(function ($id) { return (string)$id; }, $ids);
 
         return array_values(array_intersect_key($this->identityMap[$class], $ids));
     }